Moved WebViewMochaTests to shared-tests module
