--- platforms/android/bridge-webview/src/androidTest/AndroidManifest.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.salesforce.nimbus.bridge.webview">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-
-    <application
-        android:label="@string/app_name">
-        <activity android:name=".WebViewActivity"/>
-    </application>
-
-</manifest>
