--- platforms/android/bridge-webview/src/androidTest/res/layout/activity_web_view.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<WebView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/webView"
-    android:layout_width="0dp"
-    android:layout_height="0dp"
-    />
